requirements.txt
asgiref==3.10.0
Django==5.2.7
django-extensions==4.1
djangorestframework==3.16.1
environs==14.3.0
marshmallow==4.0.1
python-dotenv==1.1.1
sqlparse==0.5.3
tzdata==2025.2
